refactor: make more aliases to be lazy
